feat: переносит код в файл utils
