+++ js/main.js
@@ -1,3 +1,7 @@
+import './data.js';
+import './functions.js';
+import './utils.js';
+
 const NICKNAME_COMMENT = [
   'Кекс',
   'Мартин',
@@ -35,29 +39,6 @@ const MAX_ID_COMMENTS = 999;
 
 const MIN_ID_AVATAR = 1;
 const MAX_ID_AVATAR = 6;
-
-function getRandomInteger(a, b) {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-function createUniqNumber(minNum, maxNum) {
-  const previousNumb = [];
-
-  return function () {
-    let currentNum = getRandomInteger(minNum, maxNum);
-    while (previousNumb.includes(currentNum)) {
-      if (previousNumb.length >= (minNum - maxNum + 1)) {
-        return null;
-      }
-      currentNum = getRandomInteger(minNum, maxNum);
-    }
-    previousNumb.push(currentNum);
-    return currentNum;
-  };
-}
 
 function createPhotoDesc() {
   const randomId = createUniqNumber(MIN_ID_PHOTO, MAX_ID_PHOTO);
